# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenClinic — Electronic Health Record (EHR) & Governance API",
    "description": "Open, modular, and auditable Electronic Health Record (EHR) and Clinical Management API with hybrid access control (RBAC + ACL), multi-tenancy, dynamic capabilities resolution, and enterprise security (Argon2id + in-memory JWT).",
    "version": "1.0.0",
    "contact": {
      "name": "Iniciativa OpenClinic",
      "url": "https://github.com/Iniciativa-OpenClinic/OpenClinic"
    },
    "license": {
      "name": "GNU AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token in Authorization header: `Bearer <token>`"
      }
    },
    "schemas": {
...
      "def-0": {
        "type": "object",
        "description": "Estrutura padronizada de erros da aplicação no padrão RFC 7807",
        "properties": {
          "type": {
            "type": "string",
            "description": "URI identificadora do tipo de erro"
          },
          "title": {
            "type": "string",
            "description": "Resumo legível do erro"
          },
          "status": {
            "type": "integer",
            "description": "Código de status HTTP"
          },
          "detail": {
            "type": "string",
            "description": "Descrição detalhada e traduzida do erro"
          },
          "code": {
            "type": "string",
            "description": "Código unívoco de erro da aplicação (ex: ERR_AUTH_INVALID_CREDENTIALS)"
          },
          "instance": {
            "type": "string",
            "description": "URI do endpoint onde ocorreu o erro"
          },
          "invalid_params": {
            "type": "array",
            "description": "Lista de parâmetros com falha de validação",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "reason": {
                  "type": "string"
                }
              }
            }
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemDetails"
      }
    }
  },
  "paths": {
...
    "/api/v1/arch/tenants": {
      "get": {
        "summary": "List All Registered Tenants",
        "tags": [
          "Architecture & Multi-Tenancy"
        ],
        "description": "Returns the list of all tenants registered in sys_tenants ordered by is_default DESC, name ASC.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of registered tenants",
            "content": {
              "application/json": {
                "schema": {
                  "description": "List of registered tenants",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string"
                      },
                      "slug": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "ACTIVE",
                          "SUSPENDED",
                          "DELETED"
                        ]
                      },
                      "taxId": {
                        "type": "string",
                        "nullable": true
                      },
                      "cnpj": {
                        "type": "string",
                        "nullable": true
                      },
                      "contactName": {
                        "type": "string",
                        "nullable": true
                      },
                      "contactTitle": {
                        "type": "string",
                        "nullable": true
                      },
                      "contactEmail": {
                        "type": "string",
                        "nullable": true
                      },
                      "contactPhone": {
                        "type": "string",
                        "nullable": true
                      },
                      "postalCode": {
                        "type": "string",
                        "nullable": true
                      },
                      "street": {
                        "type": "string",
                        "nullable": true
                      },
                      "number": {
                        "type": "string",
                        "nullable": true
                      },
                      "complement": {
                        "type": "string",
                        "nullable": true
                      },
                      "neighborhood": {
                        "type": "string",
                        "nullable": true
                      },
                      "city": {
                        "type": "string",
                        "nullable": true
                      },
                      "state": {
                        "type": "string",
                        "nullable": true
                      },
                      "country": {
                        "type": "string",
                        "nullable": true
                      },
                      "isDefault": {
                        "type": "boolean"
                      },
                      "isActive": {
                        "type": "boolean"
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida ou erro de validação (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado ou token inválido/expirado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado — Permissão ou Role insuficiente (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create New Tenant",
        "tags": [
          "Architecture & Multi-Tenancy"
        ],
        "description": "Creates and registers a new organization tenant in the multi-tenant registry.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "slug"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 255
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "SUSPENDED",
                      "DELETED"
                    ],
                    "default": "ACTIVE"
                  },
                  "taxId": {
                    "type": "string",
                    "nullable": true
                  },
                  "cnpj": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 18
                  },
                  "contactName": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 255
                  },
                  "contactTitle": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 100
                  },
                  "contactEmail": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 255
                  },
                  "contactPhone": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 20
                  },
                  "postalCode": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 20
                  },
                  "street": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 255
                  },
                  "number": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 20
                  },
                  "complement": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 100
                  },
                  "neighborhood": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 100
                  },
                  "city": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 100
                  },
                  "state": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 100
                  },
                  "country": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 50
                  },
                  "isDefault": {
                    "type": "boolean",
                    "default": false
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Tenant created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Tenant created successfully",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "CREATED"
                    },
                    "message": {
                      "type": "string",
                      "example": "Tenant created successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "ACTIVE",
                            "SUSPENDED",
                            "DELETED"
                          ]
                        },
                        "taxId": {
                          "type": "string",
                          "nullable": true
                        },
                        "cnpj": {
                          "type": "string",
                          "nullable": true
                        },
                        "contactName": {
                          "type": "string",
                          "nullable": true
                        },
                        "contactTitle": {
                          "type": "string",
                          "nullable": true
                        },
                        "contactEmail": {
                          "type": "string",
                          "nullable": true
                        },
                        "contactPhone": {
                          "type": "string",
                          "nullable": true
                        },
                        "postalCode": {
                          "type": "string",
                          "nullable": true
                        },
                        "street": {
                          "type": "string",
                          "nullable": true
                        },
                        "number": {
                          "type": "string",
                          "nullable": true
                        },
                        "complement": {
                          "type": "string",
                          "nullable": true
                        },
                        "neighborhood": {
                          "type": "string",
                          "nullable": true
                        },
                        "city": {
                          "type": "string",
                          "nullable": true
                        },
                        "state": {
                          "type": "string",
                          "nullable": true
                        },
                        "country": {
                          "type": "string",
                          "nullable": true
                        },
                        "isDefault": {
                          "type": "boolean"
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida ou erro de validação (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado ou token inválido/expirado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado — Permissão ou Role insuficiente (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
}